+++ src/main/java/TeamGoat/TripSupporter/Controller/Review/ReviewController.java
@@ -12,12 +12,14 @@ import org.springframework.data.web.PagedResourcesAssembler;
 import org.springframework.hateoas.EntityModel;
 import org.springframework.hateoas.PagedModel;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/reviews")
 @Slf4j
+@Validated
 public class ReviewController {
 
     private final ReviewServiceImpl reviewService;
