+++ src/main/java/TeamGoat/TripSupporter/Exception/User/IllegalEmailFormatException.java
@@ -1,0 +1,7 @@
+package TeamGoat.TripSupporter.Exception.User;
+
+public class IllegalEmailFormatException extends RuntimeException {
+    public IllegalEmailFormatException(String message) {
+        super(message);
+    }
+}
