--- src/main/java/TeamGoat/TripSupporter/Exception/User/WrongEmailFormatException.java
@@ -1,7 +1,0 @@
-package TeamGoat.TripSupporter.Exception.User;
-
-public class WrongEmailFormatException extends RuntimeException {
-    public WrongEmailFormatException(String message) {
-        super(message);
-    }
-}
